requirements.txt
streamlit>=1.24.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.2
matplotlib>=3.7.0
Pillow>=9.0.0
fpdf>=2.7.0
python-dotenv>=1.0.0